Add powercell seeking functionality
There are two methods so far, simple (which has been tested and works to some extent), and regular (uses a PID to turn; hasn't been tested ye).

Occasionally a nullPointerException will be thrown by the getXOffset method in the PixyCam subsystem because there is an inconsistency in object detection. Either the detection must be tuned in PixyMon to avoid flickering, or the getXOffset method must return the previous offset until it is able to get a reading.